refactor: extract Chat.svelte logic into modular service modules
